Start working on strategy interface
