+++ spamalot.go
@@ -87,6 +87,15 @@ type Spammer struct {
 	metricRelay chan<- Metric
 }
 
+func (s *Spammer) NewWorker(node Node) *Worker {
+	return &Worker{
+		node:       node,
+		api:        giota.NewAPI(node.URL, nil),
+		spammer:    s,
+		stopSignal: s.stopSignal,
+	}
+}
+
 type Option func(*Spammer) error
 
 func New(options ...Option) (*Spammer, error) {
@@ -292,7 +301,7 @@ func (s *Spammer) Start() {
 	}
 
 	for _, node := range s.nodes {
-		w := worker{
+		w := Worker{
 			node:       node,
 			api:        giota.NewAPI(node.URL, nil),
 			spammer:    s,
@@ -357,201 +366,6 @@ func (s *Spammer) Start() {
 	}
 }
 
-type worker struct {
-	node       Node
-	api        *giota.API
-	spammer    *Spammer
-	stopSignal chan struct{}
-}
-
-// retrieves tips from the given node and puts them into the tips channel
-func (w worker) getNonZeroTips(tipsChan chan Tips, wg *sync.WaitGroup) {
-	defer wg.Done()
-	for {
-		select {
-		case <-w.stopSignal:
-			return
-		default:
-			tips, err := w.api.GetTips()
-			if err != nil {
-				w.spammer.logIfVerbose("GetTips error", err)
-				continue
-			}
-
-			// Loop through returned tips and get a random txn
-			// if txn value is zero, get a new one
-			var txn *giota.Transaction
-			var txnHash giota.Trytes
-			for {
-				if len(tips.Hashes) == 0 {
-					break
-				}
-
-				r := rand.Intn(len(tips.Hashes))
-				txns, err := w.api.GetTrytes([]giota.Trytes{tips.Hashes[r]})
-				if err != nil {
-					w.spammer.logIfVerbose("GetTrytes error:", err)
-					continue
-				}
-
-				txn = &txns.Trytes[0]
-				if txn.Value == 0 {
-					tips.Hashes = append(tips.Hashes[:r],
-						tips.Hashes[r+1:]...)
-					continue
-				}
-				txnHash = tips.Hashes[r]
-				break
-			}
-
-			if txn == nil {
-				continue
-			}
-
-			w.spammer.logIfVerbose("Got tips from", w.node.URL)
-
-			nodeInfo, err := w.api.GetNodeInfo()
-			if err != nil {
-				w.spammer.logIfVerbose("GetNodeInfo error:", err)
-				continue
-			}
-			txns, err := w.api.GetTrytes([]giota.Trytes{nodeInfo.LatestMilestone})
-			if err != nil {
-				w.spammer.logIfVerbose("GetTrytes error:", err)
-				continue
-			}
-
-			milestone := txns.Trytes[0]
-
-			tip := Tips{
-				Trunk:      milestone,
-				TrunkHash:  nodeInfo.LatestMilestone,
-				Branch:     *txn,
-				BranchHash: txnHash,
-			}
-
-			tipsChan <- tip
-
-		}
-	}
-}
-
-// retrieves tips from the given node and puts them into the tips channel
-func (w worker) getTxnsToApprove(tipsChan chan Tips, wg *sync.WaitGroup) {
-	defer wg.Done()
-	for {
-		select {
-		case <-w.stopSignal:
-			return
-		default:
-			tips, err := w.api.GetTransactionsToApprove(w.spammer.depth, giota.NumberOfWalks, "")
-			if err != nil {
-				w.spammer.logIfVerbose("GetTransactionsToApprove error", err)
-				continue
-			}
-
-			txns, err := w.api.GetTrytes([]giota.Trytes{
-				tips.TrunkTransaction,
-				tips.BranchTransaction,
-			})
-
-			if err != nil {
-				//return nil, err
-				w.spammer.logIfVerbose("GetTrytes error:", err)
-				continue
-			}
-
-			w.spammer.logIfVerbose("Got tips from", w.node.URL)
-
-			tip := Tips{
-				Trunk:      txns.Trytes[0],
-				TrunkHash:  tips.TrunkTransaction,
-				Branch:     txns.Trytes[1],
-				BranchHash: tips.BranchTransaction,
-			}
-
-			tipsChan <- tip
-
-		}
-	}
-}
-
-// receives prepared txs and attaches them via remote node or local PoW onto the tangle
-func (w worker) spam(txnChan <-chan Transaction, wg *sync.WaitGroup) {
-	defer wg.Done()
-	for {
-
-		select {
-		case <-w.stopSignal:
-			return
-			// read next tx to processes
-		case txn, ok := <-txnChan:
-			if !ok {
-				return
-			}
-
-			switch {
-			case !w.spammer.localPoW && w.node.AttachToTangle:
-
-				w.spammer.logIfVerbose("attaching to tangle")
-
-				at := giota.AttachToTangleRequest{
-					TrunkTransaction:   txn.Trunk,
-					BranchTransaction:  txn.Branch,
-					MinWeightMagnitude: w.spammer.mwm,
-					Trytes:             txn.Transactions,
-				}
-
-				attached, err := w.api.AttachToTangle(&at)
-				if err != nil {
-					w.spammer.metrics.addMetric(INC_FAILED_TX, nil)
-					log.Println("Error attaching to tangle:", err)
-					continue
-				}
-
-				txn.Transactions = attached.Trytes
-			default:
-
-				// lock so only one worker is doing PoW at a time
-				w.spammer.powMu.Lock()
-				w.spammer.logIfVerbose("doing PoW")
-
-				err := doPow(&txn, w.spammer.depth, txn.Transactions, w.spammer.mwm, w.spammer.pow)
-				if err != nil {
-					w.spammer.metrics.addMetric(INC_FAILED_TX, nil)
-					log.Println("Error doing PoW:", err)
-					w.spammer.powMu.Unlock()
-					continue
-				}
-				w.spammer.powMu.Unlock()
-			}
-
-			err := w.api.BroadcastTransactions(txn.Transactions)
-
-			w.spammer.RLock()
-			defer w.spammer.RUnlock()
-			if err != nil {
-				w.spammer.metrics.addMetric(INC_FAILED_TX, nil)
-				log.Println(w.node, "ERROR:", err)
-				continue
-			}
-
-			// this will auto print metrics to console
-			w.spammer.metrics.addMetric(INC_SUCCESSFUL_TX, txandnode{txn, w.node})
-
-			// wait the cooldown before accepting a new TX
-			if w.spammer.cooldown > 0 {
-				select {
-				case <-w.stopSignal:
-					return
-				case <-time.After(w.spammer.cooldown):
-				}
-			}
-		}
-
-	}
-}
-
 func (s *Spammer) Stop() error {
 	// nil the tip and txs channel so that send/receive
 	// on those channels becomes blocking
@@ -572,11 +386,6 @@ func (s *Spammer) IsRunning() bool {
 	s.RLock()
 	defer s.RUnlock()
 	return s.running
-}
-
-type Tips struct {
-	Trunk, Branch         giota.Transaction
-	TrunkHash, BranchHash giota.Trytes
 }
 
 type Transaction struct {
@@ -667,36 +476,4 @@ func (s *Spammer) buildTransactions(trytes []giota.Transaction, pow giota.PowFun
 		}, nil
 	}
 
-}
-
-func doPow(tra *Transaction, depth int64, trytes []giota.Transaction, mwm int64, pow giota.PowFunc) error {
-	var prev giota.Trytes
-	var err error
-	for i := len(trytes) - 1; i >= 0; i-- {
-		switch {
-		case i == len(trytes)-1:
-			trytes[i].TrunkTransaction = tra.Trunk
-			trytes[i].BranchTransaction = tra.Branch
-		default:
-			trytes[i].TrunkTransaction = prev
-			trytes[i].BranchTransaction = tra.Trunk
-		}
-
-		trytes[i].Nonce, err = pow(trytes[i].Trytes(), int(mwm))
-		if err != nil {
-			return err
-		}
-
-		prev = trytes[i].Hash()
-	}
-	return nil
-}
-
-func padTag(tag string) string {
-	for {
-		tag += "9"
-		if len(tag) > 27 {
-			return tag[0:27]
-		}
-	}
 }
