+++ util.go
@@ -1,0 +1,35 @@
+package spamalot
+
+import "github.com/CWarner818/giota"
+
+func doPow(tra *Transaction, depth int64, trytes []giota.Transaction, mwm int64, pow giota.PowFunc) error {
+	var prev giota.Trytes
+	var err error
+	for i := len(trytes) - 1; i >= 0; i-- {
+		switch {
+		case i == len(trytes)-1:
+			trytes[i].TrunkTransaction = tra.Trunk
+			trytes[i].BranchTransaction = tra.Branch
+		default:
+			trytes[i].TrunkTransaction = prev
+			trytes[i].BranchTransaction = tra.Trunk
+		}
+
+		trytes[i].Nonce, err = pow(trytes[i].Trytes(), int(mwm))
+		if err != nil {
+			return err
+		}
+
+		prev = trytes[i].Hash()
+	}
+	return nil
+}
+
+func padTag(tag string) string {
+	for {
+		tag += "9"
+		if len(tag) > 27 {
+			return tag[0:27]
+		}
+	}
+}
